backend/requirements.txt
astapi
uvicorn[standard]
python-multipart
pillow
groq
python-dotenv
pydantic